Admin controller is in under work

+++ src/main/java/org/backend/Controllers/AdminController.java
@@ -1,14 +1,17 @@
 package org.backend.Controllers;
 
+import org.backend.DTOs.HikeRouteErrorDTO;
+import org.backend.DTOs.HikeRouteSuccessDTO;
+import org.backend.DTOs.ResponseDTO;
 import org.backend.Model.HikeRoute;
 import org.backend.Repository.HikeRouteRepository;
-import org.backend.Model.Message;
 import org.backend.Service.HikeRouteService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-
 import java.util.List;
+import java.util.Optional;
 
+@CrossOrigin(origins = "*", allowedHeaders = "*")
 @RestController
 public class AdminController {
     HikeRouteService hikeRouteService;
@@ -23,17 +26,27 @@ public class AdminController {
 
     }
     @DeleteMapping("/hike_routes/{hikeRouteId}")
-    public String deleteHikeRoute(@PathVariable long hikeRouteId){
-        hikeRouteRepository.deleteById(hikeRouteId);
-        return "Törlés sikeres!";
+    public ResponseDTO deleteHikeRoute(@PathVariable long hikeRouteId){
+        Optional<HikeRoute> hikeRoute=hikeRouteRepository.findById(hikeRouteId);
+        if(hikeRoute.isPresent()){
+            hikeRouteRepository.deleteById(hikeRouteId);
+            return new HikeRouteSuccessDTO();
+
+        }else {
+            return new HikeRouteErrorDTO();
+        }
+
 
     }
     @GetMapping("/hike_routes")
     public List<HikeRoute> getAllHikeRoute(){
         return hikeRouteRepository.findAll();
     }
-    @PutMapping("/hike_routes/{hikeRouteId}/{mesaage}")
-    public String modifyHikeRoute(@PathVariable Long hikeRoutId,@PathVariable Message message){
-    return "Módosítás sikeres!";
+
+    @PutMapping("/hike_routes")
+    public List<HikeRoute> modifyHikeRoute(@RequestBody HikeRoute hikeRoute ){
+        hikeRouteRepository.save(hikeRoute);
+    return hikeRouteRepository.findAll();
     }
+
 }
